+++ NEWS.md
@@ -7,6 +7,8 @@ Default values are aligned with Kliche et al. 2014.
 (`get_region_matrix()`) #5.
 * Method `nchars()` is implemented for `subcorpus` and `subcorpus_bundle` objects 
 now and will be available for `plpr_subcorpus` by inheritance #3.
+* An outdated example für the `Duplicates` class using a `duplicates()` method
+has been removed #7.
 
 ## v0.1.9